pyclipr stub
